refactor: change module structure
