--- src/app/features/movie/models/ui/movie-list.model.ts
@@ -1,5 +1,0 @@
-import { MovieCard } from './movie-card.model';
-
-export interface MovieList {
-  movies: MovieCard[];
-}
