+++ src/app/features/movie/movie-card/movie-card.model.ts
