+++ src/app/features/movie/movie-list/movie-list.model.ts
@@ -1,0 +1,5 @@
+import { MovieCard } from '../movie-card/movie-card.model';
+
+export interface MovieList {
+  movies: MovieCard[];
+}
