+++ src/app/features/movie/movies.module.ts
@@ -1,71 +1,15 @@
-import { Component, model } from '@angular/core';
-import { MovieListComponent } from '@features/movie/components/movie-list/movie-list.component';
-import { MovieList } from '@features/movie/models/ui/movie-list.model';
-import { MoviesService } from '@features/movie/services/movies.service';
-import { MovieSearchComponent } from '@features/movie/components/movie-search/movie-search.components';
-import { CommonModule, NgFor } from '@angular/common';
-import { environment } from '@environments/environment.development';
-import { MovieCard } from '@features/movie/models/ui/movie-card.model';
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+import MoviesComponent from './movies.component';
 
-@Component({
-  selector: 'movies',
-  templateUrl: './components/movies/movies.component.html',
-  standalone: true,
-  imports: [MovieListComponent, MovieSearchComponent, NgFor, CommonModule],
-  host: {
-    class: 'w-full',
-  },
+@NgModule({
+  imports: [
+    RouterModule.forChild([
+      {
+        path: '',
+        component: MoviesComponent,
+      },
+    ]),
+  ],
 })
-export default class MoviesModule {
-  movieLists = model<MovieList | undefined>();
-  currentMovie = model<MovieCard | undefined>();
-  loadingState = true;
-
-  constructor(private movieService: MoviesService) {
-    this.movieService
-      .get({
-        sort_by: 'vote_count.desc',
-        'vote_average.lte': 10,
-      })
-      .subscribe((movies) => {
-        const movieArr: MovieCard[] = movies.results.map((movie, idx) => {
-          return {
-            id: idx + 1,
-            title: movie.title,
-            img: environment.imageUrl + movie.poster_path,
-            score: Math.round(movie.vote_average * 100) / 100,
-            backdrop: environment.imageUrl + movie.backdrop_path,
-            overview: movie.overview,
-            release_date: movie.release_date,
-          };
-        });
-        this.movieLists.set({ movies: movieArr });
-        this.currentMovie.set(movieArr[0]);
-        this.loadingState = false;
-      });
-  }
-
-  search(value: string) {
-    this.movieService.search({ query: value }).subscribe((movies) => {
-      this.loadingState = true;
-      const movieArr: MovieCard[] = movies.results.map((movie, idx) => {
-        return {
-          id: idx + 1,
-          title: movie.title,
-          img: environment.imageUrl + movie.poster_path,
-          score: Math.round(movie.vote_average * 100) / 100,
-          backdrop: environment.imageUrl + movie.backdrop_path,
-          overview: movie.overview,
-          release_date: movie.release_date,
-        };
-      });
-      this.movieLists.set({ movies: movieArr });
-      this.currentMovie.set(movieArr[0]);
-      this.loadingState = false;
-    });
-  }
-
-  selectMovie(movie: MovieCard) {
-    this.currentMovie.set(movie);
-  }
-}
+export class MoviesModule {}
